Further cleanup DataTableComponent - move method
